MACVENTURE: Fix cursor warning
