+++ engines/macventure/image.cpp
@@ -24,7 +24,7 @@
 
 namespace MacVenture {
 
-PPICHuff PPIC1Huff = {
+static const PPICHuff PPIC1Huff = {
 	// Masks
 	{ 0x0000,0x2000,0x4000,0x5000,0x6000,0x7000,0x8000,0x9000,0xa000,
 	0xb000,0xc000,0xd000,0xd800,0xe000,0xe800,0xf000,0xf800 },
@@ -35,7 +35,7 @@ PPICHuff PPIC1Huff = {
 	0x02,0x04,0x0b,0x0d,0xe }
 };
 
-PPICHuff PPIC2Huff = {
+static const PPICHuff PPIC2Huff = {
 	// Masks
 	{ 0x0000,0x4000,0x8000,0xc000,0xc800,0xd000,0xd800,0xe000,0xe800,
 	0xf000,0xf400,0xf600,0xf800,0xfa00,0xfc00,0xfe00,0xff00 },
@@ -47,18 +47,19 @@ PPICHuff PPIC2Huff = {
 };
 
 // Used to load the huffman table in PPIC3 decoding
-byte loadBits[] = {
-		0x08, 0x0f, 0x02, 0xff, 0x00,
-		0x04, 0xff, 0x01,
-		0x07, 0x09, 0x08, 0xff, 0x03,
-		0x04, 0xff, 0x04,
-		0x0a, 0x07, 0x0a, 0x0b, 0x06, 0xff, 0x05,
-		0x06, 0x06, 0x0b, 0xff, 0x07,
-		0x03, 0xff, 0x09,
-		0x04, 0x03, 0x0e, 0xff, 0x0c,
-		0x02, 0xff, 0x0d,
-		0x01, 0xff, 0x0f,
-		0xff };
+static const byte loadBits[] = {
+	0x08, 0x0f, 0x02, 0xff, 0x00,
+	0x04, 0xff, 0x01,
+	0x07, 0x09, 0x08, 0xff, 0x03,
+	0x04, 0xff, 0x04,
+	0x0a, 0x07, 0x0a, 0x0b, 0x06, 0xff, 0x05,
+	0x06, 0x06, 0x0b, 0xff, 0x07,
+	0x03, 0xff, 0x09,
+	0x04, 0x03, 0x0e, 0xff, 0x0c,
+	0x02, 0xff, 0x0d,
+	0x01, 0xff, 0x0f,
+	0xff
+};
 
 ImageAsset::ImageAsset(ObjID original, Container * container) {
 	_id = (original * 2);
